--- mlflow/exceptions.py
@@ -1,6 +1,0 @@
-class MlflowException(Exception):
-    """Base exception in MLflow."""
-
-
-class IllegalArtifactPathError(MlflowException):
-    """The artifact_path parameter was invalid."""
