+++ mlflow/exceptions/__init__.py
@@ -1,1 +1,10 @@
-from mlflow.utils.exception import 
+from mlflow.projects.exceptions import ExecutionException
+from mlflow.data import DownloadException
+from mlflow.store.rest_store import RestException
+from mlflow.utils.process import ShellCommandException
+from mlflow.store.dbfs_artifact_repo import IllegalArtifactPathError
+
+
+__all__ = [
+    "ExecutionException", "DownloadException", "RestException", "ShellCommandException",
+    "IllegalArtifactPathError"]
