+++ mlflow/projects/exceptions.py
@@ -1,7 +1,3 @@
-class MLflowException(Exception):
-    pass
-
-
 class ExecutionException(Exception):
     """Exception thrown when executing a project fails."""
     pass
