+++ mlflow/store/artifact_repo.py
@@ -1,6 +1,5 @@
 from abc import abstractmethod, ABCMeta
 
-from mlflow.exceptions import MlflowException
 from mlflow.store.rest_store import DatabricksStore
 
 
@@ -83,7 +82,7 @@ class ArtifactRepository:
         elif artifact_uri.startswith("dbfs:/"):
             from mlflow.store.dbfs_artifact_repo import DbfsArtifactRepository
             if not isinstance(store, DatabricksStore):
-                raise MlflowException('`store` must be an instance of DatabricksStore.')
+                raise TypeError('`store` must be an instance of DatabricksStore.')
             return DbfsArtifactRepository(artifact_uri, store.http_request_kwargs)
         else:
             from mlflow.store.local_artifact_repo import LocalArtifactRepository
