+++ mlflow/store/dbfs_artifact_repo.py
@@ -2,7 +2,6 @@ import json
 import os
 
 from mlflow.entities.file_info import FileInfo
-from mlflow.exceptions import IllegalArtifactPathError, MlflowException
 from mlflow.store.artifact_repo import ArtifactRepository
 from mlflow.utils.file_utils import build_path, get_relative_path, TempDir
 from mlflow.utils.rest_utils import http_request, RESOURCE_DOES_NOT_EXIST
@@ -11,6 +10,10 @@ from mlflow.utils.string_utils import strip_prefix
 LIST_API_ENDPOINT = '/api/2.0/dbfs/list'
 GET_STATUS_ENDPOINT = '/api/2.0/dbfs/get-status'
 DOWNLOAD_CHUNK_SIZE = 1024
+
+
+class IllegalArtifactPathError(Exception):
+    """The artifact_path parameter was invalid."""
 
 
 def _dbfs_list_api(json, http_request_kwargs):
@@ -62,7 +65,7 @@ class DbfsArtifactRepository(ArtifactRepository):
         cleaned_artifact_uri = artifact_uri.rstrip('/')
         super(DbfsArtifactRepository, self).__init__(cleaned_artifact_uri)
         if not cleaned_artifact_uri.startswith('dbfs:/'):
-            raise MlflowException('DbfsArtifactRepository URI must start with dbfs:/')
+            raise ValueError('DbfsArtifactRepository URI must start with dbfs:/')
         self.http_request_kwargs = http_request_kwargs
 
     def _get_dbfs_path(self, artifact_path):
