+++ tests/projects/test_entry_point.py
@@ -5,7 +5,7 @@ import pytest
 from six.moves import shlex_quote
 
 
-from mlflow.utils.exception import ExecutionException
+from mlflow.projects import ExecutionException
 from mlflow.utils.file_utils import TempDir
 from tests.projects.utils import load_project, TEST_PROJECT_DIR
